// apps/user-experience/components/Cookie/Cookie.type.ts
export interface CookieComponentProps {
  data: CookieDataType;
}

export type CookieDataType = {
  cookie_policy_cta_link: string;
  non_essential_cookie_description: string;
  cookie_manage_setting_consent_button: string;
  cookie_button_text: string;
  non_essential_cookie_title: string;
  manage_save_setting_consent_button: string;
  essential_cookie_description: string;
  consent_cookie_button: string;
  manage_setting_description: string;
  consent_cookie_title: string;
  manage_setting_cookie_button: string;
  cookie_description: string;
  cookie_manage_setting_title: string;
  essential_cookie_title: string;
  consent_cookie_description: string;
  cookie_policy_cta_text: string;
  cookie_title: string;
  consent_cookie_policy_link: string;
  informative_cookie_policy_link: string;
  consent_cookie_country_list: string;
  informative_cookie_country_list: string;
  cookie_consent_expiry_time: string;
  cookie_informative_expiry_time: string;
};